requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0

# Model and training
timm>=0.6.0
scikit-learn>=1.0.0

# Data processing
pandas>=1.3.0
Pillow>=8.3.0
requests>=2.25.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional dependencies for advanced features
# wandb>=0.12.0  # For experiment tracking (uncomment if needed)
# umap-learn>=0.5.0  # For UMAP visualization (uncomment if needed)

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0